requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.3
SQLAlchemy==1.4.41
flask-jwt-extended
Flask-Migrate
flask-cors
Pillow
Faker
Flask-Seeder
openai
python-dotenv